Move method to correct library
